routing refactor - I. add && search crashes
