+++ server.js
@@ -3,12 +3,10 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const app = express();
 const router = express.Router();
-const add = require('./src/server').add;
-const removeItem = require('./src/server').removeItem;
-const updateItem = require('./src/server').updateItem;
-const searchItem = require('./src/server').searchItem;
-const getById = require('./src/server').getById;
 const database = require('./src/server/database');
+
+//FIXME: remove this
+const add = require('./src/server').add;
 
 //config
 const PORT = 4125;
@@ -27,7 +25,7 @@ database
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
-// middleware to use for all requests
+// middleware
 router.use(function(req, res, next) {
   if (LAST_ID < 0) {
     res.status(503).json({ message: 'server not connected to DB, please wait for a while' });
@@ -41,6 +39,8 @@ router.use(function(req, res, next) {
 router.get('/', function(req, res) {
   res.json({ message: 'hooray! welcome to our api!' });
 });
+
+//UPLOAD
 router.route('/movies/upload').post(function(req, res) {
   for (let i in req.body.movie) {
     let current = req.body.movie[i];
@@ -50,69 +50,20 @@ router.route('/movies/upload').post(function(req, res) {
   }
   res.json({ message: 'ok' }); //FIXME: server sends 'that correct' but its not truth
 });
-//ADD
-router
-  .route('/movies')
-  .post(function(req, res) {
-    LAST_ID++;
-    let movie = { ...req.body.movie, id: LAST_ID };
-    let success = function() {
-      res.json({ message: movie, status: 'ok' });
-    };
-    add(success, database, movie, LAST_ID);
-  });
-
-//SEARCH
-//FIXME move search to server method, add postman scheme for search
-router.route('/movies/search').get(function(req, res) {
-  let success = function(i) {
-    res.json({ message: i, status: 'ok' });
-  };
-  searchItem(success, database, req.url);
-});
-
-//GET BY ID
-router
-  .route('/movies/:movieId')
-  .get(function(req, res) {
-    let id = req.params.movieId;
-    let successHandler = function(i) {
-      res.json({message:i});
-    };
-    let failHandler = function () {
-      res.status(426).json({ message: 'item not found' });
-    };
-    getById(successHandler, database, id, failHandler);
-  })
-  //UPDATE
-  .put(function(req, res) {
-    let id = req.params.movieId;
-    let movie = req.body.movie;
-    let success = function() {
-      res.json({ message: req.body.movie, status: 'ok' });
-    };
-    updateItem(success, database, id, movie);
-  })
-  //DELETE
-  .delete(function(req, res) {
-    let id = req.params.movieId;
-    let success = function(i) {
-      res.json({ message: req.body.movie, status: 'ok' });
-    };
-    let fail = function(err) {
-      res.status(426).json({ message: err });
-    };
-    let notFound = function() {
-      console.log('not found called');
-      res.status(426).json({ message: 'item not found' });
-    };
-
-    removeItem(success, notFound, fail, database, id);
-  });
 
 app.use('/api', require('./src/server/getAll'));
+app.use('/api', require('./src/server/searchItem'));
+app.use('/api', require('./src/server/addItem'));
+
+app.use('/api', require('./src/server/getById'));
+app.use('/api', require('./src/server/updateItem'));
+app.use('/api', require('./src/server/deleteItem'));
+
 app.use('/api', router);
 
 const server = app.listen(PORT, function() {
   console.log('app running on port.', server.address().port);
 });
+
+//FIXME: VERY UGLY
+module.exports.LAST_ID = LAST_ID;
