+++ src/server/getAll.js
@@ -9,7 +9,7 @@ router.get('/movies', function(req, res) {
     .then(function(snapshot) {
       res.json(snapshot.val().movie);
     })
-  //FIXME #2 add error handling for requests
+    //FIXME #2 add error handling for requests
     .catch();
 });
